apps/backend/services/notifications/requirements.txt
# Core Framework
fastapi
uvicorn[standard]

# Messaging
kafka-python

# Utilities
python-dotenv
python-json-logger
Jinja2
